Batch experimental design update
