+++ tests/emukit/experimental_design/test_batch_experimental_design.py
@@ -11,7 +11,7 @@ from emukit.core.acquisition import Acquisition
 from emukit.core.interfaces import IModel
 from emukit.core.loop.loop_state import create_loop_state
 from emukit.core.optimization import AcquisitionOptimizer
-from emukit.experimental_design.model_based.batch_experimental_design import GreedyBatchPointCalculator
+from emukit.core.loop.candidate_point_calculators import GreedyBatchPointCalculator
 from emukit.experimental_design.model_based.experimental_design_loop import ExperimentalDesignLoop
